Revisão geral do livro #72

+++ livro/capitulos/code/cap2/reg_conta.c
@@ -1,5 +1,5 @@
 typedef struct {
-	long    numero_da_conta;
-	long    cpf_do_cliente;
-	double  saldo;
+	long       numero_da_conta;
+	long long  cpf_do_cliente;
+	double     saldo;
 } Conta;
